requirements.txt
Django>=4.0,<5.0
psycopg2-binary>=2.9
djangorestframework>=3.12
django-cors-headers>=3.10
django-filter>=2.4
djangorestframework-simplejwt>=4.7
pandas>=1.3
numpy>=1.21
reliability>=0.4.0
matplotlib>=3.4
scipy>=1.7
openpyxl>=3.0
requests==2.31.0 
django-unfold==0.65.0